+++ packages/e2e-tests/src/bundler.ts
@@ -46,6 +46,7 @@ export async function startBundler(config: BundlerConfig): Promise<void> {
   console.log("Writing to bundler env file:\n", envFile);
   fs.writeFileSync(BUNDLER_ENV_FILE_PATH, envFile);
   await compose.upAll(BUNDLER_COMPOSE_OPTS);
+
   await sleep(3_000);
 }
 
@@ -54,4 +55,5 @@ export async function stopBundler(): Promise<void> {
     cwd: BUNDLER_COMPOSE_CWD,
     commandOptions: [["--volumes"]],
   });
+  await sleep(3_000);
 }